Move train to loops
